GENIE/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113 
torchvision==0.12.0+cu113
click==8.1.3
numpy==1.22.4
pillow==9.2.0
tqdm==4.62.2
opencv-python==4.6.0.66
lmdb==1.3.0
scipy==1.7.1
matplotlib==3.6.1
absl-py==1.3.0
ml_collections==0.1.1
ninja==1.10.2